torch.equal checks if two tensors are equal in both size and values. For
named tensors, it also checks that the names are exactly equal. There is
an argument to be made for alternative semantics (check that the names
*match*), but for an API that is called "equal" I would expect it to
check equality on names as well.

Test Plan:
- new tests

[ghstack-poisoned]

+++ aten/src/ATen/TensorNames.h
@@ -41,7 +41,7 @@ struct CAFFE2_API TensorName {
  private:
   ArrayRef<Dimname> origin_;
   Dimname name_;
-  uint8_t origin_idx_; // A named tensor can have at most 64 dims.
+  int origin_idx_; // A named tensor can have at most 64 dims.
 
   CAFFE2_API friend std::ostream& operator<<(
       std::ostream& out,
